requirements.txt
pillow
flask_socketio
flask-cors==3.0.10
beautifulsoup4==4.12.2
bs4==0.0.1
lxml==4.9.2
pandas
requests==2.28.2
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
pygtail